test: Remove unused txmempool include from tests
