+++ src/Makefile.test_util.include
@@ -5,20 +5,21 @@
 LIBTEST_UTIL=libtest_util.a
 
 EXTRA_LIBRARIES += \
-    $(LIBTEST_UTIL)
+  $(LIBTEST_UTIL)
 
 TEST_UTIL_H = \
-    test/util/blockfilter.h \
-    test/util/chainstate.h \
-    test/util/logging.h \
-    test/util/mining.h \
-    test/util/net.h \
-    test/util/script.h \
-    test/util/setup_common.h \
-    test/util/str.h \
-    test/util/transaction_utils.h \
-    test/util/validation.h \
-    test/util/wallet.h
+  test/util/blockfilter.h \
+  test/util/chainstate.h \
+  test/util/logging.h \
+  test/util/mining.h \
+  test/util/net.h \
+  test/util/script.h \
+  test/util/setup_common.h \
+  test/util/str.h \
+  test/util/transaction_utils.h \
+  test/util/txmempool.h \
+  test/util/validation.h \
+  test/util/wallet.h
 
 libtest_util_a_CPPFLAGS = $(AM_CPPFLAGS) $(BITCOIN_INCLUDES) $(BOOST_CPPFLAGS)
 libtest_util_a_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)
@@ -31,6 +32,7 @@ libtest_util_a_SOURCES = \
   test/util/setup_common.cpp \
   test/util/str.cpp \
   test/util/transaction_utils.cpp \
+  test/util/txmempool.cpp \
   test/util/validation.cpp \
   test/util/wallet.cpp \
   $(TEST_UTIL_H)
