+++ src/test/blockencodings_tests.cpp
@@ -7,6 +7,7 @@
 #include <consensus/merkle.h>
 #include <pow.h>
 #include <streams.h>
+#include <test/util/txmempool.h>
 
 #include <test/util/setup_common.h>
